Two sizes taken from the hardware instead of from me
MAXVISPLANES is now overridable, and still 128. The guard exists so a build
can be SCAFFOLDED -- lowered just far enough to boot a map whose limits are
not yet known, so the watermarks can be read -- and DOOM_SCAFFOLD in
build_arm.sh is the deliberately awkward way to pass that, because
scaffolding that persists is an undocumented cut. The measurement stays in
the comment beside it: 93 of 128 across all 36 Freedoom maps, and 29 of them
need more than 16.

viewangletox is short, not int. R_InitTextureMapping clamps every entry to
[-1, viewwidth+1], so 321 is the widest value it can hold, and saying so is
8,192 bytes of RAM.

Both came out of getting Doom to run on a Pico 2 W and then asking it what
it actually used, rather than reasoning about it:

  stack      1,984 of 32,768 reserved
  heap peak  256,448

The stack was 32,768 because I picked that number. It is 8,192 now, which
is still four times what the hardware reports, and the 24,576 it gave back
is most of what the port was short.

+++ doomgeneric/r_main.c
@@ -90,7 +90,10 @@ angle_t			clipangle;
 // maps the visible view angles to screen X coordinates,
 // flattening the arc to a flat projection plane.
 // There will be many angles mapped to the same X. 
-int			viewangletox[FINEANGLES/2];
+/* short, not int: R_InitTextureMapping clamps every entry to [-1,
+ * viewwidth+1], so the widest value this can hold is 321 and the array is
+ * 8,192 bytes of RAM smaller for saying so. */
+short			viewangletox[FINEANGLES/2];
 
 // The xtoviewangleangle[] table maps a screen pixel
 // to the lowest viewangle that maps back to x ranges
